Renamed and moved ContentView Changes I renamed the main ContentView file to RecipeListView since this will still serve as the main screen of the app but is now referenced as something more relevant. I created a /Views directory and then moved the renamed file into this directory. I also switched from using #Preview {} to use a PreviewProvider.
Future Directions
I am going to push this commit to set the remote origin for the dev branch and then I will set up the /Services directory and create the function to access the first api endpoint to GET the list of dessert recipes which I will print out so that I can see exactly what information I can access with this first api call. Then I will make sure I am removing null entries and values and possibly duplicates as well just in case. After this I will set up the model for the RecipeListItems or call it something similar and this file will be placed in a /Models directory.

Note
N/A