Constants for URLs Changes Created a new directory /Utilities and then created a new file called Constants.swift which contain all of the constants used in the project. The only things I have so far are the baseURL for the api and then the two endpoints. The Dessert category filter endpoint is complete as there is nothing to be added onto this from the app but the endpoint for fetching details leaves space for the id of the recipe to be concatenated on afterwards in the api call function.
Future Directions
I will set up the api service for getting the recipes and then will print this to the console so I can see what information I can get and use when rendering the List. Then I will set up the model for the list items and then I will create the ViewModel and then finally the View.

Note
N/A