Recipe Model Changes Created Recipe.swift inside of the new /Models directory which is the model for the items that will be fed as an array to the List element on the RecipeListView. This model is identifiable using the isMeal field of the json data and then also codable enumerating the respective fields from the json data to the fields created for the model.
Future Directions
I need to update the api service function so that it handles the mapping of the json data to the keys that are in the model and also returns error messages instead of just printing everything. I am stil going to have it print the output. I also need to make functions to clean the data of null or empty values and possibly duplicates and the add sorting to sort the list in alphabetical order based off of the first character in strMeal/name. Then I might create a separate component which will render a single recipe and then I will put everything together and render the list of the compnents.

Note
N/A