fetchRecipes Updated and printed Changes I updated the fetchRecipes function in APIService.swift so that it uses the model and decodes the json data from the api response to the Recipe model and then prints the array out now. I added some additional guards to check for http errors because I was getting some weird warnings in the console but from what I can see in the logs which I have many of currently it appears that everything is functioning exactly like it should be so I am not too worried about it at the moment. I will revisit this later when everything is all said and done.
Future Directions
I am going to create a component which will be what is rendered as each item in the RecipeListView. After I create this component then I will end up setting up the RecipeListView with the data the way it is currently and then I will move all of the logic to a new file which is the ViewModel and then I will create new functions like removing null or missing values, removing duplicates if there are any and then sorting the list alphabetically off of the first character in the name field.

Note
N/A