Recipe Changes Created a new file in /Views called Recipe which is simply what a single recipe would look like in the RecipeList. I have a preview of this view and it looks acceptable for now.
Future Directions
Calling it a night and I will be back at it tomorrow. Going to push the changes. Tomorrow I set up the RecipeListView with all of the logic on the RecipeListView file and then I will migrate it all over to a ViewModel file and then I will add the functions to remove nulls and empty values, remove duplicates, and sort in alphabetical order.

Note
I currently have a placeholder as ProgressView but maybe I actually set something up in the Assets and have a default image url or simply remove the recipe altogether if it is missing data. Not sure what would be best.