RecipeListView functioning Changes Set up some state variables to hold onto the array of recipes and another for errorMessages. I modified the function to the fetch the recipes on this file so that it handles populating the array and errors if needed. Then I set up the list to render the RecipeView component I made as the items and then set it up so that when the screen mounts the recipes are fetched and the list is populated.
Future Directions
Going to start the /ViewModels directory and then move all of the fetching logic from the RecipeListView to this viewmodel and then make sure everything works. Aftet the migration I am going to add the functions one at a time to clean the data and then sort it by alphabetical order.

Note
UI is terrible at the moment. At the least I need to remove the spacers on the outside of the RecipeView and just have a fixed horizontal padding which will keep everything to the sides of the view. Then I need to set a fixed max width for the text element and then do a left align on the text. I think I will move the text to it will be on the left side of the view and image on the right and then the left align will look nice and it should clean things up a good bit.