RecipeListViewModel migration Changes Created /ViewModels directory and then new file RecipeListViewModel which will handle all of the logic to call the fetchRecipes service function, process and sort the data. Currently migrated the fetching logic to the ViewModel file and updated the RecipeListView to utilize the ViewModel.
Future Directions
Next plan is to create a new function on the ViewModel that will remove null and empty values. Then a function to remove duplicates and then finally a function to sort the array of recipes. After all of that is done then I plan on setting up the function to make the next api request and print that to the console.

Note
Could be an issue with multiple renders which could be due to the image states changing or something else. Will investigate this further soon.