Filter null & empty items Changes I made name and thumbnailURL optional in the model so that nothing breaks if these values are null or empty when assembling the array from json data. I then created the function to remove recipes if those fields are missing in the data. This was followed by incorporating this function into the data being used by the ViewModel.This and the other functions will be called in a controller function which is the one that is used by the ViewModel. Since the fields are optional now I had to modify the way the View for each list item is set up so that a default is set.
Future Directions
Function to remove duplicates in the case that multiple of the same recipe get returned by the server somehow. After this I will create a function to sort everything alphabetically. Then I will set up the api service function to get the recipe details and print this to the console.

Note
N/A