fetchRecipeDetails Changes New function in api services file to fetch the recipe details from id and print the output to the console. I assigned this function as the action on a button on the main view to test with a predetermined id number.
Future Directions
Set up a model to handle the data that is provided by the api then update the api request function to use this model then assign the function as an action whenever a list item is selected and still print the output. Then create a new view for the recipedetails and add some navigation.

Note
N/A