RecipeDetailsView Changes Create new file in /Views called RecipeDetailsView which will be what is rendered when a user selects a recipe from the list. I current have fetching logic on here.Looks terrible at the moment but is working with example input.
Future Directions
Going to set up all of the navigation so that the user can get from one view to the next and then pass the id from the selected item to the view so that the data can be fetched. After this is all functioning I will make the ViewModel for this and then migrate the logic over. Then I will set up the data cleaning and prep before rendering. If all goes well here then I can make everything look pretty and then set up all the tests.

Note
Data needs a lot more cleaning than anticipated.