Data Cleaning Changes Added function to the viewmodel to remove empty or null values and then update the viewmodel to use this function. Also realized there was an issue initializing some of the fields in my model so I fixed that as well.
Future Directions
Might add a function to make an array from the string instructions based off of newlines so that I can display these steps better. Then I am going to redo the layout of the RecipeDetailsView because it looks terrible. Then the custom backbutton and navigation titles and then add a splash of color, add tests, CI flow, icon and splash screen then push and merge and then README.

Note
Still weird behavior of navigation link so need to see what to do about that