Async/Await Changes Created some files to have unit tests for the view models and api service respectively and realized that using async await would be significantly better for all of my data fetching. So I reworked the functions in APIService and then updated both viewModels to use these new asynchronous functions and then updated the RecipeListView and RecipeDetailsView to use task instead of onAppear. Everything seems to be working just fine.
Future Directions
Finish the unit tests, then UI tests, then the CI/CD workflow with github actions. If all of that goes well and doesn't take too long I'm going to whip up a quick icon and splash screen. Then push to remote, open pull request to merge w main and hopefully the tests pass and I can merge and do a quick writeup.

Note
I want to get rid of the little bar on the side of the scrollview so I'm going to see if I can't remove that quickly.