RecipeListViewModelTests Changes set up unit tests for all non-data-fetching functions from the RecipeListViewModeland all are passing currently. I noticed that there was an issue with the sorting function which was treating capitalization differently so I changed that function and now the sorting looks better. I started setting up the tests for non-data-fetching functions from the RecipeDetailsViewModel but not quite done yet.
Future Directions
Finish the unit tests, start UI tests.

Note
I think I might add a capitalization function to the RecipeListViewModel for namesso that the first character of each word in a recipe name is capitalized when it is rendered. I also will use the function i already have on the RecipeDetailsViewModel in a similar fashion.